+++ docs/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 * :bug:`-` ENS avatar fetching for each name's avatar should now work for all name resolvers. Even older or custom ones.
 * :bug:`-` Users will now be able to decode compounding transactions for Convex gauges.
 * :bug:`-` Users will no longer see double counting of balances in curve gauges.
+* :bug:`-` Aave transactions that deposit ETH in a contract that uses WETH will now be correctly decoded.
 * :bug:`-` Events that were imported via a csv file will now be properly accounted for in PnL reports.
 * :feature:`-` The PnL report page will load faster if there is many old reports in the DB.
 * :feature:`-` Convex staking and Curve gauge balances will no longer need a manual balances referesh to be detected.